Sam White from Univ of IL provided the fix for multiple CoarseScaleModels tp run.
From Sam:
I found the issue and a workaround. The problem is that on PE 0, Charm++ creates chare array elements inline at the ckNew call, but for remote chare creation, a message must be sent. In the constructor for the CoarseScaleModel  you are creating CkCallbacks for reductions using the global readonly variable 'coarseScaleArray', which the same as 'thisProxy' but is not set at the time of chare array construction on PE 0.

If you use 'thisProxy' in place of the global readonly variable 'coarseScaleArray' in the coarseScaleModel constructor when creating the CkCallbacks, your code works for all 'counts' of CoarseScaleModel and numbers of PEs. We are now considering how to address this semantic difference between PE 0 and the rest of the PEs in Charm++'s chare array construction.

The change has been made and multiple CoarseScaleModels can now run.
The results are currently different for different numbers of CoarseScaleModels, they should be the same.